requirements.txt
black==22.12.0
Flask==2.2.2
paho-mqtt==1.6.1
pytest==7.2.1
